+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New locales:
   - Swedish
   - Russian
+- New config option `catalogTitleAfterImage` to set a different title in the header after a logo
 
 ### Changed
 
@@ -22,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Disable temporal extent filter when a single date/time is provided as temporal extent in the Collection metadata
 - Better default STAC title detection within not fully loaded lists where only a URL is available
 - No search / sort functionality available when a static catalog has only a subset of children loaded
+- The default value for `catalogTitle` is `null` instead of `STAC Browser`.
 
 ### Fixed